refactor(lua): extract ValueType<->codec conversions into value_conv.rs
Stage 3 of #89. Moves virtual_value_from_type/typed_value_from_type/value_to_type and
their int/float helpers out of lua.rs, separating the value conversion from the sim
runtime. lua.rs 833->709.
